+++ platforms/core-configuration/configuration-cache/src/crossVersionTest/groovy/org/gradle/internal/cc/impl/ConfigurationCacheProblemsCrossVersionTest.groovy
@@ -26,7 +26,7 @@ import org.gradle.tooling.events.ProgressListener
 import org.gradle.tooling.events.problems.LineInFileLocation
 import org.gradle.tooling.events.problems.Severity
 import org.gradle.tooling.events.problems.SingleProblemEvent
-import org.gradle.tooling.events.problems.internal.GeneralData
+import org.gradle.tooling.events.problems.internal.DefaultAdditionalData
 
 @ToolingApiVersion(">=8.10 <8.12")
 @TargetGradleVersion(">=8.10")
@@ -83,7 +83,7 @@ class ConfigurationCacheProblemsCrossVersionTest extends ToolingApiSpecification
             definition.severity == Severity.ERROR
             (locations[0] as LineInFileLocation).path == "build file 'build.gradle'" // FIXME: the path should not contain a prefix nor extra quotes
             (locations[1] as LineInFileLocation).path == "build file '$buildFile.path'"
-            additionalData instanceof GeneralData
+            additionalData instanceof DefaultAdditionalData
             additionalData.asMap.isEmpty()
         }
     }
