+++ platforms/core-configuration/configuration-cache/src/crossVersionTest/groovy/org/gradle/internal/cc/impl/ConfigurationCacheProblemsTapi813PlusCrossVersionTest.groovy
@@ -27,7 +27,7 @@ import org.gradle.tooling.events.problems.LineInFileLocation
 import org.gradle.tooling.events.problems.Problem
 import org.gradle.tooling.events.problems.Severity
 import org.gradle.tooling.events.problems.SingleProblemEvent
-import org.gradle.tooling.events.problems.internal.GeneralData
+import org.gradle.tooling.events.problems.internal.DefaultAdditionalData
 
 @ToolingApiVersion(">=8.13")
 @TargetGradleVersion(">=8.10")
@@ -84,7 +84,7 @@ class ConfigurationCacheProblemsTapi813PlusCrossVersionTest extends ToolingApiSp
             definition.severity == Severity.ERROR
             (originLocations[0] as LineInFileLocation).path == "build file 'build.gradle'" // FIXME: the path should not contain a prefix nor extra quotes
             (originLocations[1] as LineInFileLocation).path == "build file '$buildFile.path'"
-            additionalData instanceof GeneralData
+            additionalData instanceof DefaultAdditionalData
             additionalData.asMap.isEmpty()
         }
     }
