+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/events/problems/Problem.java
@@ -17,6 +17,7 @@
 package org.gradle.tooling.events.problems;
 
 
+import org.gradle.api.Action;
 import org.gradle.api.Incubating;
 import org.gradle.tooling.Failure;
 
@@ -94,6 +95,14 @@ public interface Problem {
 
     /**
      * Returns the additional data associated with this problem.
+     *
+     * There are 2 possible types for additional data:
+     * <ul>
+     *     <li>{@link CustomAdditionalData} - custom additional data attached to a problem by using {@link org.gradle.api.problems.ProblemSpec#additionalData(Class, Action)}.
+     *     <li>{@link AdditionalData} - additional data attached to the problem.
+     * </ul>
+     *
+     * to determine the type of additional data {@code instanceof} can be used.
      *
      * @return the additional data
      * @since 8.12
